% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikhonov.R
\name{cornerScurve}
\alias{cornerScurve}
\title{Find corner S-curve}
\usage{
cornerScurve(A, beta, b, lambda, z, verbose = FALSE)
}
\arguments{
\item{A}{the A matrix of the system: Ax = b.}

\item{beta}{regression coefficients.}

\item{b}{the b vector of the system: Ax = b.}

\item{lambda}{grid of penalty values.}

\item{z}{intermediate matrix for L-curvature calculation.}

\item{verbose}{Verbose TRUE/FALSE}
}
\value{
index for the corner of the S-curve.
}
\description{
Find corner S-curve
}
\details{
details follow.
}
\author{
Maarten van Iterson
}
