% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution}
\alias{deconvolution}
\title{deconvolution estimator using fft}
\usage{
deconvolution(object)
}
\arguments{
\item{object}{of class SampleSize}
}
\value{
object of class SampleSize
}
\description{
TODO: suppress warning message in pt() :  full precision was not achieved in 'pnt'
TODO:replace massdist by trimbin?
deconvolution estimator using fft
}
\details{
details follow
}
\author{
Maarten van Iterson
}
