\name{deepSAGE}
\alias{deepSAGE}
\docType{data}
\title{Test statistics derived from a deepSAGE experiment}
\description{follow  
}
\usage{data(deepSAGE)}
\format{
  A vector of 44882 test statistics.
  \describe{Vector of test statistics obtained by performing a likelihood ratio test using edgeR
  }
}
\details{follow
}
\source{
't Hoen, P.A.C. Ariyurek, Y. Thygesen, H.H. Vreugdenhil, E. Vossen, R.H.A.M. de Menezes, R.X. Boer, J.M. van Ommen, G.B. and den Dunnen, J.T.,
Deep Sequencing-based Expression analysis shows Major Advances in Robustness, Resolution and Inter-lab Portability over Five Microarray Platforms,
Nucleic Acids Research, 2008.
}
%\references{}
\examples{
data(deepSAGE)
str(deepSAGE)
}
\keyword{datasets}
