% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSizeClass.R
\name{defineEffectSizeRange}
\alias{defineEffectSizeRange}
\title{defineEffectSizeRange}
\usage{
defineEffectSizeRange(object, from, to, resolution)
}
\arguments{
\item{object}{pilotdata object}

\item{from}{lower limit effect-sizes}

\item{to}{upper limit effect-sizes}

\item{resolution}{should be power of 2 default length statistics round to nearest power of 2}
}
\value{
effect-size range
TODO use trimmingbinning function for this!
}
\description{
defineEffectSizeRange
}
