% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerandsamplesize.R
\name{findroot}
\alias{findroot}
\title{bisection method for finding ustar}
\usage{
findroot(g, h, umax, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{g}{left-hand-side}

\item{h}{right-hand side}

\item{umax}{between [0, 1] usually alpha.}

\item{verbose}{TRUE/FALSE}

\item{plot}{TRUE/FALSE}
}
\value{
ustar
}
\description{
bisection method for finding ustar
}
\details{
solve G(u) = u pi0 (1-alpha)/(aplha(1-pi0)), more  details follow.
}
\author{
Maarten van Iterson
}
