% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericalintegration.R
\name{midpoint}
\alias{midpoint}
\title{Midpoint rule for numerical integration.}
\usage{
midpoint(f, a, b, n, ...)
}
\arguments{
\item{f}{bivariate function.}

\item{a}{lower bound of the grid.}

\item{b}{upper bound of the grid.}

\item{n}{grid size.}

\item{...}{trick for evaluating the second parameter in case a bivariate functions was supplied.}
}
\value{
vector or matrix of function evaluations use sum to obtain the integrand.
}
\description{
Implementation of the midpoint rule for the numerical integration of uni- and bivariate functions.
}
\details{
details follow
}
\author{
Maarten van Iterson
}
