% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugategradient.R
\name{nncg}
\alias{nncg}
\title{Non-negative conjugate gradient algorithm}
\usage{
nncg(A, b, type = 1, trace = FALSE)
}
\arguments{
\item{A}{the A matrix of the system: Ax = b.}

\item{b}{the b vector of the system: Ax = b.}

\item{type}{for the conjugate-gradients method. Takes value '1' for the Fletcher-Reeves update, '2' for Polak-Ribiere and '3' for Beale-Sorenson.}

\item{trace}{tracing information on the progress of the optimization is produced.}
}
\value{
list containg regression coefficients and some additional information.
}
\description{
Non-negative conjugate gradient algorithm
}
\details{
C-implementation, details follow.
}
\author{
Maarten van Iterson
}
