% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikhonov.R
\name{picardplot}
\alias{picardplot}
\title{Picard-plot}
\usage{
picardplot(
  A,
  b,
  xlim = c(1, length(b)),
  ylim = NULL,
  main = "Picard-Plot",
  legend = TRUE
)
}
\arguments{
\item{A}{the A matrix of the system: Ax = b.}

\item{b}{the b vector of the system: Ax = b.}

\item{xlim}{xlim of Picard-plot.}

\item{ylim}{ylim of Picard-plot.}

\item{main}{main}

\item{legend}{legend}
}
\value{
generates Picard-plot.
}
\description{
Generates Picard-plot
}
\details{
details follow.
}
\author{
Maarten van Iterson
}
