\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,ANY-method}
\alias{plot,PilotData-method}
\alias{plot,SampleSize-method}
\title{Methods for Function \code{plot} in Package \pkg{SSPA}}
\description{
 Plot function for objects of class PilotData and SampleSize
}
\section{Methods}{
\describe{

\item{\code{signature(x = "PilotData")}}{
 Diagonstic plots of the PilotData.
}

\item{\code{signature(x = "SampleSize")}}{
 Plot the estimated density of effect sizes.
}
}}
\keyword{methods}
