\name{predictpower}
\alias{predictpower}
\title{Predict power for given vector of sample sizes}
\usage{
  predictpower(object, samplesizes, alpha = 0.1,
    verbose = FALSE, plot = FALSE)
}
\arguments{
  \item{object}{of class 'SampleSize'}

  \item{samplesizes}{vector of total sample sizes.}

  \item{alpha}{FDR.}

  \item{verbose}{TRUE/FALSE}

  \item{plot}{TRUE/FALSE}
}
\value{
  predicted power.
}
\description{
  Predict power for given vector of sample sizes
}
\details{
  details follow.
}
\author{
  Maarten van Iterson
}

