% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikhonov.R
\name{regularization}
\alias{regularization}
\title{Find optimal regularization parameter}
\usage{
regularization(
  b,
  A,
  beta,
  edf,
  lambda,
  z,
  method = c("lcurve", "gcv", "aic"),
  plot = TRUE,
  log = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{b}{the b vector of the system: Ax = b.}

\item{A}{the A matrix of the system: Ax = b.}

\item{beta}{regression coefficients.}

\item{edf}{effective degrees of freedom.}

\item{lambda}{grid of penalty values.}

\item{z}{intermediate matrix for L-curvature calculation.}

\item{method}{Either the L-curve, GCV or AIC.}

\item{plot}{Plot TRUE/FALSE.}

\item{log}{Plot on log-scale TRUE/FALSE.}

\item{verbose}{Verbose TRUE/FALSE}
}
\value{
generates optionally figure and returns the index for the optimal penalty value.
}
\description{
Find optimal regularization parameter
}
\details{
details follow.
}
\author{
Maarten van Iterson
}
