\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ANY-method}
\alias{show,PilotData-method}
\alias{show,SampleSize-method}
\title{General show method for Classes PilotData and SampleSize}
\description{
Methods for function \code{show} in package \pkg{SSPA}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "PilotData")}}{
Show the content of a PilotData-object in a userfriendly way.
}

\item{\code{signature(object = "SampleSize")}}{
Show the content of a SampleSize-object in a userfriendly way.
}

}}
\keyword{methods}
