\name{show-methods}
\docType{methods}
\alias{show,PilotData-method}
\alias{show,SampleSize-method}

\title{Nicely showing objects of class pilotData and sampleSize}

\description{Nicely showing objects from \code{"\linkS4class{PilotData}"} and \code{"\linkS4class{SampleSize}"}}

\section{Methods}{
\describe{
\item{object = "PilotData"}{object from class \code{"\linkS4class{PilotData}"}}
\item{object = "SampleSize"}{object from class \code{"\linkS4class{SampleSize}"}}
}}

\keyword{methods}

