% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikhonov.R
\name{standardform}
\alias{standardform}
\title{Transform to standard from}
\usage{
standardform(b, A, L)
}
\arguments{
\item{b}{the b vector of the system: Ax = b.}

\item{A}{the A matrix of the system: Ax = b.}

\item{L}{non-diagonal penalty}
}
\value{
Transformed system ||Ax - b || + lambda ||L|| with L diagonal.
}
\description{
Transforms an objective function containing non-diagonal penalty to standard form (diagonal penalty)
}
\details{
details follow.
}
\author{
Maarten van Iterson
}
