% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmingbinning.R
\name{trimbin}
\alias{trimbin}
\title{Binning and trimming of the test statistics vector.}
\usage{
trimbin(
  statistics,
  nbins = 100,
  trim = c(0.01, 0.99),
  bin = c("epdf", "ecdf"),
  symmetric = TRUE,
  plot = TRUE
)
}
\arguments{
\item{statistics}{vector of test statistics.}

\item{nbins}{number of bins.}

\item{trim}{vector of size two with lower and upper limits for the trimming.}

\item{bin}{"epdf" or "ecdf" binning using respectively, empirical density of cumulutive distribution function.}

\item{symmetric}{symmetric trimming TRUE/FALSE.}

\item{plot}{plot results TRUE/FALSE.}
}
\value{
trimmed and binned vector of test statistics.
}
\description{
Binning and trimming of the test statistics vector.
}
\details{
details follow
}
\author{
Maarten van Iterson
}
